requirements.txt
   streamlit
   opencv-python-headless
   tensorflow
   keras
   fer
   openai-whisper
   ffmpeg-python
   moviepy
   textblob
   